+++ FrontEnd/src/CSS/Login.css
@@ -29,7 +29,11 @@ img{
 }
 
 #form{
-    margin-left: 20px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 50%;
 }
 
 #textFields{
@@ -75,13 +79,19 @@ img{
     color: red;
 }
 
-#loginButton{
-    margin: auto;
-    width: 30%;
-    border: 2px solid #0F5530;
+#microsoftButton{
+    text-decoration: none;
+    border: 2px #0F5530 solid;
+    border-radius: 2px;
     padding: 10px;
-    margin-top: 20px;
-    font-size: 15px;
+    color: #0F5530;
+    width: 60%;
+    font-size: 18px;
+    text-align: center;
+}
+
+#microsoftButton:hover{
+    background-color: lightgray;
 }
 
 #successfulLogin{
